ai-trading-dashboard/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
ta==0.11.0
statsmodels==0.14.2
transformers==4.41.2
torch==2.3.0
newsapi-python==0.2.7
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
joblib==1.4.2
tensorflow==2.16.1
keras==3.3.3
scipy==1.13.0
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
requests==2.32.2
aiohttp==3.9.5
python-multipart==0.0.9
